+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 2.5 - (something)
 ## 2.5.1
 - Updated `permissions` command
+- Added `greeting` command
 
 ## 2.5.0 (btw, this is probably the biggest update ever)
 - Fixed `rank, ranking, math` commands